Update server to 1.7.10
This commit upgrades the server to 1.7.10. This also means that the
EULA must be accepted for the server to start up. This can be done
manually by editing the `eula.txt` (ew, manually), or by setting the
`node['minecraft']['accept_eula']` attribute to true. We can't
automatically accept EULAs for users, so this is false by default.

* Add attribute for accepting the EULA
* Write out a eula.txt resource with the content based on whether the
  EULA was accepted
* Some minor readme cleanup